Backup/requirements.txt
django-allauth
celery
redis
django-apscheduler
python-dotenv
django-celery-beat
django-celery-results # currently requires Django 4.0 or older
python-decouple
pillow
sorl-thumbnail
mkdocs





